Make some TPLs optional and user settable
What this commit allows is to completely ignore detection and attempting
to build a given TPL by the user. This is done by passing a cmake
argument, e.g.,

-DENABLE_DOXYGEN=false

which will ignore detecing and attempting to find Doxygen on the system.

This is useful if a user only wants a partial build, e.g., only wants to
build some of the executables of quinoa or if a TPL is not needed, e.g.,
in a docker image which will never want use, e.g., Doxygen, because it
will never attempt to build target 'doc'.

The TPLs now optional are:

HDF5 (default on),
NetCDF (default: on),
Boost (default: on),
Hypre (default: on),
TestU01 (default: on),
Trilinos (default: on),
Doxygen, (default: on),
m.css, (default: on),
Root, (default: off), (was already settable before this),
BackwardCpp (default: on), (was already settable before this),
Omega_H, (default: off), (was already settable before this).