Switch alpin/ampi build to a dynamic one
This commit removes setting the BUILD_SHARED_LIBBS flag as true from the
alpine/ampi build. As a result the hdf5 patch that removes get_time()
from hdf5 is not applied, which appears to be required for a static
build for the official hdf5 but breaks hdf5-ampi. Thus this is a
work-around that switches the alpine tpl build to a shared-object build
instead of tryeing to deal with the problem explicitly within Charm++'s
fork of hdf5.